Añadido header en todos los archivos

+++ frontend/javascript/coordinator-campaigns.js
@@ -4,11 +4,21 @@ document.addEventListener('DOMContentLoaded', async () => {
     const token = localStorage.getItem('token');
     if (!token) { window.location.href = 'login.html'; return; }
 
-    document.getElementById('user-name').textContent = localStorage.getItem('nombre') || 'Coordinador';
-    document.getElementById('btn-logout').addEventListener('click', () => {
-        localStorage.clear();
-        window.location.href = 'login.html';
-    });
+    const userNameEl = document.getElementById('user-name');
+	
+    if (userNameEl) {
+        userNameEl.textContent = localStorage.getItem('nombre') || 'Coordinador';
+    }
+
+    document.addEventListener('click', (e) => {
+        if(e.target.id === 'btn-edit'){
+            window.location.href = 'edit.html';
+            
+        } else if(e.target.id === 'btn-logout'){
+            localStorage.clear();
+            window.location.href = 'login.html';
+        }
+    })
 
     const tbody = document.getElementById('campaigns-tbody');
